backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
qdrant-client==1.6.4
sentence-transformers==2.2.2
reportlab==4.0.7
qrcode==7.4.2
pypdf2==3.0.1
python-pptx==0.6.23
python-docx==1.1.0
openai==1.3.7
httpx==0.25.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
